+++ addon/mixins/flexberry-objectlistview-route.js
@@ -44,43 +44,32 @@ export default Mixin.create({
       let saveBeforeRouteLeave = methodOptions.saveBeforeRouteLeave;
       let onEditForm = methodOptions.onEditForm;
       let editFormRoute = methodOptions.editFormRoute;
+      let recordId = record.get('id') || record.get('data.id');
+      let thisRouteName = this.get('router.currentRouteName');
+      let thisRecordId = this.get('currentModel.id');
       let transitionOptions = {
         queryParams: {
           modelName: methodOptions.modelName,
-          customParameters:  methodOptions.customParameters
+          customParameters:  methodOptions.customParameters,
+          parentRoute: thisRouteName,
+          parentRouteRecordId: thisRecordId
         }
       };
       if (!editFormRoute) {
         throw new Error('Detail\'s edit form route is undefined.');
       }
 
-      let recordId = record.get('id') || record.get('data.id');
-      let thisRouteName = this.get('router.currentRouteName');
-      let thisRecordId = this.get('currentModel.id');
       if (!onEditForm) {
-        this.transitionTo(editFormRoute, recordId, transitionOptions)
-        .then((newRoute) => {
-          if (newRoute) {
-            newRoute.controller.set('parentRoute', thisRouteName);
-          }
-        });
+        this.transitionTo(editFormRoute, recordId, transitionOptions);
       } else {
         if (saveBeforeRouteLeave) {
           this.controller.save(false, true).then(() => {
-            this.transitionTo(editFormRoute, recordId, transitionOptions)
-            .then((newRoute) => {
-              newRoute.controller.set('parentRoute', thisRouteName);
-              newRoute.controller.set('parentRouteRecordId', thisRecordId);
-            });
+            this.transitionTo(editFormRoute, recordId, transitionOptions);
           }).catch((errorData) => {
             this.controller.rejectError(errorData, this.get('i18n').t('forms.edit-form.save-failed-message'));
           });
         } else {
-          this.transitionTo(editFormRoute, recordId, transitionOptions)
-          .then((newRoute) => {
-            newRoute.controller.set('parentRoute', thisRouteName);
-            newRoute.controller.set('parentRouteRecordId', thisRecordId);
-          });
+          this.transitionTo(editFormRoute, recordId, transitionOptions);
         }
       }
     },