+++ tests/acceptance/components/flexberry-objectlistview/folv-user-button-test.js
@@ -1,5 +1,4 @@
 import { executeTest } from './execute-folv-test';
-import $ from 'jquery';
 
 executeTest('user button test', (store, assert, app) => {
   assert.expect(5);
@@ -10,15 +9,17 @@ executeTest('user button test', (store, assert, app) => {
     assert.equal(currentPath(), path);
 
     let controller = app.__container__.lookup('controller:' + currentRouteName());
-    let $testBudtton = $('.test-click-button')[0];
+
+    // Enable the hi button.
+    click('.toggle-hi-button');
 
     // First click.
-    $testBudtton.click();
-    assert.equal(controller.clickCounter, 2, 'Test button was pressed');
+    click('.test-click-button');
+    andThen(() => assert.equal(controller.clickCounter, 2, 'Test button was pressed'));
 
     // Second click.
-    $testBudtton.click();
-    assert.equal(controller.clickCounter, 3, 'Test button was pressed');
+    click('.test-click-button');
+    andThen(() => assert.equal(controller.clickCounter, 3, 'Test button was pressed'));
 
     assert.notOk(controller.get('modelFromClickedRow'));
     click('.ui.button > .bug.icon:first');